refactor: ContentSection and SectionHeader to components
let's reserve the ui folder for radix
