+++ app/about/careers/page.tsx
@@ -3,7 +3,7 @@ import { Workday } from "@/components/Workday"
 import { Hero } from "@/components/Hero"
 import { getWorkdayData } from "@/app/about/queries"
 import Spinner from "@/components/assets/Spinner"
-import { ContentSection } from "@/components/ui/ContentSection"
+import { ContentSection } from "@/components/ContentSection"
 import ButtonLink from "@/components/ui/ButtonLink"
 
 export default async function AboutLayout() {
