+++ components/ContentSection.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react"
 import { cn } from "@/lib/utils"
-import { SectionHeader } from "@/components/ui/SectionHeader"
+import { SectionHeader } from "@/components/SectionHeader"
 
 interface ContentSectionProps {
   children: ReactNode
