+++ components/SectionHeader.tsx
